+++ pytket/binders/passes.cpp
@@ -526,18 +526,9 @@ PYBIND11_MODULE(passes, m) {
       "SynthesiseTket", &SynthesiseTket,
       "Optimises and converts all gates to CX, TK1 and Phase gates.");
   m.def(
-      "SynthesiseOQC",
-      []() {
-        tket_log()->warn(
-            "SynthesiseOQC is deprecated. It will be removed "
-            "after pytket v1.28.");
-        return SynthesiseOQC();
-      },
-      "Optimises and converts all gates to ECR, Rz, SX and Phase. "
-      "DEPRECATED: will be removed after pytket 1.28.");
-  m.def(
       "SynthesiseUMD", &SynthesiseUMD,
-      "Optimises and converts all gates to XXPhase, PhasedX, Rz and Phase.");
+      "Optimises and converts all gates to XXPhase, PhasedX, Rz and Phase. "
+      "DEPRECATED: will be removed after pytket 1.32.");
   m.def(
       "SquashTK1", &SquashTK1,
       "Squash sequences of single-qubit gates to TK1 gates.");